Implement passive maimai fights

+++ server/UserData/Log/claim_log.cs
@@ -1,9 +1,9 @@
 namespace Nanina.UserData.Log;
 
-public class ClaimLog(Game game, float xp, float difficultyRating)
+public class ClaimLog(Game game, float xp, double difficultyRating)
 {
     public Game game { get; set; } = game;
     public float xp { get; set; } = xp;
     public long timestamp { get; set; } = Utils.GetTimestamp();
-    public float difficultyRating { get; set; } = difficultyRating;
+    public double difficultyRating { get; set; } = difficultyRating;
 }
